requirements.txt
opencv-python>=4.5.0
numpy>=1.19.0
moviepy>=2.0.0
google-generativeai>=0.3.0
python-dotenv>=0.19.0
